Monday, December 11, 2023 at 5:42 PM West Africa Standard Time
