+++ app/components/landingPage/Hero/utils/imageSwitchHelper.js
@@ -1,0 +1,10 @@
+
+export const switchCenterImage = (
+  centerImageIndex,
+  setCenterImageIndex,
+  dynamicCenterImages
+) => {
+  setCenterImageIndex(
+    (prevIndex) => (prevIndex + 1) % dynamicCenterImages.length
+  );
+};
